Refactor: decrease uninitialized member values and the risk of memory leaks in  abacus (#5755)
* init variable and fix memory leak bug in module_base

* fix uninit member var in module_basis

* chang logic bug and uninit memeber value in the module cell

* chang logic bug and uninit memeber value in the module_elecstate

* chang logic bug and uninit memeber value in the module_eslover

* chang memory leak risk and uninit memeber value in the module_hamilt_general

* chang  uninit memeber value in the module_hamilt_lcao

* chang  uninit memeber value in the module_hamilt_pwdft

* fix in klist

* [pre-commit.ci lite] apply automatic fixes

* fix useless assert

* revert change in read_pp_upf100.cpp

* change uninit memeber value in module_hamilt_pw

* change uninit memeber value in module_hsolver

* change uninit memeber value in module_io

* change uninit memeber value in module_md and module_relax

* [pre-commit.ci lite] apply automatic fixes

* update nnr init data

* fix uninit member value

* update value in elecstate_energy.cpp

* add init parameter for construct in sto_func

* add value in the construct file

* [pre-commit.ci lite] apply automatic fixes

* change magnetism.cpp

* [pre-commit.ci lite] apply automatic fixes

* modify logic of elecstate_energy.cpp

---------

+++ source/module_cell/module_paw/paw_element.cpp
@@ -67,7 +67,7 @@ void Paw_Element::read_paw_xml(std::string filename)
         this->lstate[istate] = this->extract_int(line,"l=");
         lmax = std::max(lmax, lstate[istate]);
 
-        int pos = line.find("f=");
+        size_t pos = line.find("f=");
         if(pos!=std::string::npos)
         {
             this->lstate_occ[istate] = this->extract_double(line,"f=");
@@ -178,7 +178,7 @@ std::string Paw_Element::scan_file(std::ifstream &ifs, std::string pattern)
 
 double Paw_Element::extract_double(std::string line, std::string key)
 {
-    int index = line.find(key);
+    size_t index = line.find(key);
     if (index != std::string::npos)
     {
         std::stringstream tmp;
@@ -198,7 +198,7 @@ double Paw_Element::extract_double(std::string line, std::string key)
 
 std::string Paw_Element::extract_string(std::string line, std::string key)
 {
-    int index = line.find(key);
+    size_t index = line.find(key);
     if (index != std::string::npos)
     {
         std::stringstream tmp;
@@ -218,7 +218,7 @@ std::string Paw_Element::extract_string(std::string line, std::string key)
 
 int Paw_Element::extract_int(std::string line, std::string key)
 {
-    int index = line.find(key);
+    size_t index = line.find(key);
     if (index != std::string::npos)
     {
         std::stringstream tmp;
